Preparations for proper tests.
Move the main driver into the library.
Set up a test suite without tests.
