+++ tests/Main.hs
@@ -1,0 +1,18 @@
+module Main where
+
+-- import Simala.Main
+-- 
+-- import System.Directory
+-- import System.FilePath
+-- import System.IO.Silently
+import Test.Hspec
+-- import Test.Hspec.Golden
+
+-- import Paths_simala
+
+main :: IO ()
+main = do
+  -- dataDir <- getDataDir
+  -- exampleFiles <- getDirectoryContents (dataDir </> "examples")
+  -- print exampleFiles
+  hspec $ pure ()
